+++ src/parser/visitors/ast/ContextHandlers/HekiliApi/index.ts
@@ -1,2 +1,3 @@
 export * from "./BossHandler";
+export * from "./GroupMembersHandler";
 export * from "./SettingsHandler";